third_party/WebKit/LayoutTests/fast/canvas/offscreencanvas-text/text/2d.text.font.parse.size.percentage.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.text.font.parse.size.percentage</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/external/wpt/common/canvas-tests.js"></script>

<h1>2d.text.font.parse.size.percentage</h1>
<p class="desc"></p>


<script>
var t = async_test("");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.font = '50% serif';
_assertSame(ctx.font, '72px serif', "ctx.font", "'72px serif'");
canvas.setAttribute('style', 'font-size: 100px');
_assertSame(ctx.font, '72px serif', "ctx.font", "'72px serif'");

t.done();

});
</script>
